requirements.txt
langchain>=0.3.0
langgraph>=0.2.0
langchain-openai>=0.2.0
chromadb>=0.5.0
openai>=1.0.0
python-dotenv>=1.0.0
streamlit>=1.38.0
requests>=2.31.0
pytest>=8.0.0
